Faster way to do AoC 2019.4
Instead of looping through every int in the range, I now do something
different.

I find the first number higher than the low range that does not
decrement between the digits.

I pass this initial number into a validNumberGenerator. This removes all
"9"s from the end of the number, increments the (now smaller) number,
and appends the last digit until the length of the string is 6. If the
new number is within range, push to the output array.

This greatly reduces the initial set of valid numbers. I then loop
through this smaller set to check for duplicates. The code is much
faster now.